+++ crates/json-formatter/src/lib.rs
@@ -1,0 +1,168 @@
+use serde_json::ser::Formatter;
+use std::io;
+
+macro_rules! tri {
+	($e:expr $(,)?) => {
+		match $e {
+			core::result::Result::Ok(val) => val,
+			core::result::Result::Err(err) => return core::result::Result::Err(err),
+		}
+	};
+}
+
+/// This structure pretty prints a JSON value to make it human readable.
+#[derive(Clone, Debug)]
+pub struct JsonFormatter<'a> {
+	current_indent: usize,
+	array_breaks: bool,
+	has_value: bool,
+	indent: &'a [u8],
+}
+
+impl<'a> JsonFormatter<'a> {
+	/// Construct a pretty printer formatter that defaults to using two spaces for indentation.
+	pub fn new() -> Self {
+		JsonFormatter::with_indent(b"  ")
+	}
+
+	/// Construct a pretty printer formatter that uses the `indent` string for indentation.
+	pub fn with_indent(indent: &'a [u8]) -> Self {
+		JsonFormatter {
+			current_indent: 0,
+			array_breaks: true,
+			has_value: false,
+			indent,
+		}
+	}
+
+	/// Construct a pretty printer formatter that optionally break arrays into multiple lines.
+	pub fn with_array_breaks(array_breaks: bool) -> Self {
+		JsonFormatter {
+			current_indent: 0,
+			array_breaks,
+			has_value: false,
+			indent: b"  ",
+		}
+	}
+}
+
+impl<'a> Default for JsonFormatter<'a> {
+	fn default() -> Self {
+		JsonFormatter::new()
+	}
+}
+
+impl<'a> Formatter for JsonFormatter<'a> {
+	#[inline]
+	fn begin_array<W>(&mut self, writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.current_indent += 1;
+		self.has_value = false;
+		writer.write_all(b"[")
+	}
+
+	#[inline]
+	fn end_array<W>(&mut self, writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.current_indent -= 1;
+
+		if self.has_value && self.array_breaks {
+			tri!(writer.write_all(b"\n"));
+			tri!(indent(writer, self.current_indent, self.indent));
+		}
+
+		writer.write_all(b"]")
+	}
+
+	#[inline]
+	fn begin_array_value<W>(&mut self, writer: &mut W, first: bool) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		if !self.array_breaks {
+			if !first {
+				tri!(writer.write_all(b", "));
+			}
+
+			return Ok(());
+		}
+
+		tri!(writer.write_all(if first { b"\n" } else { b",\n" }));
+		indent(writer, self.current_indent, self.indent)
+	}
+
+	#[inline]
+	fn end_array_value<W>(&mut self, _writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.has_value = true;
+		Ok(())
+	}
+
+	#[inline]
+	fn begin_object<W>(&mut self, writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.current_indent += 1;
+		self.has_value = false;
+		writer.write_all(b"{")
+	}
+
+	#[inline]
+	fn end_object<W>(&mut self, writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.current_indent -= 1;
+
+		if self.has_value {
+			tri!(writer.write_all(b"\n"));
+			tri!(indent(writer, self.current_indent, self.indent));
+		}
+
+		writer.write_all(b"}")
+	}
+
+	#[inline]
+	fn begin_object_key<W>(&mut self, writer: &mut W, first: bool) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		tri!(writer.write_all(if first { b"\n" } else { b",\n" }));
+		indent(writer, self.current_indent, self.indent)
+	}
+
+	#[inline]
+	fn begin_object_value<W>(&mut self, writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		writer.write_all(b": ")
+	}
+
+	#[inline]
+	fn end_object_value<W>(&mut self, _writer: &mut W) -> io::Result<()>
+	where
+		W: ?Sized + io::Write,
+	{
+		self.has_value = true;
+		Ok(())
+	}
+}
+
+fn indent<W>(wr: &mut W, n: usize, s: &[u8]) -> io::Result<()>
+where
+	W: ?Sized + io::Write,
+{
+	for _ in 0..n {
+		tri!(wr.write_all(s));
+	}
+
+	Ok(())
+}