+++ src/resolution.rs
@@ -63,16 +63,14 @@ impl UnresolvedValue {
 
 			Variant::Float32(n) => AmbiguousValue::Number(n as f64),
 			Variant::Float64(n) => AmbiguousValue::Number(n),
-			Variant::Int32(n) => AmbiguousValue::Number(truncate_float(n as f64)),
+			Variant::Int32(n) => AmbiguousValue::Number(n as f64),
 			Variant::Int64(n) => AmbiguousValue::Number(n as f64),
 
 			Variant::String(str) => AmbiguousValue::String(str),
 			Variant::Tags(tags) => AmbiguousValue::StringArray(tags.iter().map(|s| s.to_string()).collect()),
 			Variant::Content(content) => AmbiguousValue::String(content.into_string()),
 
-			Variant::Vector2(vector) => {
-				AmbiguousValue::Array2([truncate_float(vector.x as f64), truncate_float(vector.y as f64)])
-			}
+			Variant::Vector2(vector) => AmbiguousValue::Array2([vector.x as f64, vector.y as f64]),
 			Variant::Vector3(vector) => AmbiguousValue::Array3([vector.x as f64, vector.y as f64, vector.z as f64]),
 			Variant::Color3(color) => AmbiguousValue::Array3([color.r as f64, color.g as f64, color.b as f64]),
 
@@ -293,12 +291,13 @@ fn list_examples(values: &[&str]) -> String {
 	output
 }
 
-// Temporary solution to avoid saving `null` values in JSON files
-fn truncate_float(float: f64) -> f64 {
-	if float.is_infinite() {
-		999_999_999.0 * float.signum()
+#[inline]
+fn truncate_number(number: &f64) -> f64 {
+	// Temporary solution to avoid saving `null` values in JSON files
+	if number.is_infinite() {
+		999_999_999.0 * number.signum()
 	} else {
-		float
+		(*number * 1_000_000.0).trunc() / 1_000_000.0
 	}
 }
 
@@ -306,7 +305,7 @@ fn serialize_number<S>(number: &f64, serializer: S) -> Result<S::Ok, S::Error>
 where
 	S: Serializer,
 {
-	let number = (*number * 1_000_000.0).trunc() / 1_000_000.0;
+	let number = truncate_number(number);
 
 	if number.fract() == 0.0 {
 		serializer.serialize_i64(number as i64)
@@ -322,7 +321,7 @@ where
 	let mut seq = serializer.serialize_seq(Some(array.len()))?;
 
 	for number in array {
-		let number = (*number * 1_000_000.0).trunc() / 1_000_000.0;
+		let number = truncate_number(number);
 
 		if number.fract() == 0.0 {
 			seq.serialize_element(&(number as i64))?;